surge, ralph: fix four defects the native build warnings were hiding (#153)
A clean `make all` emits 136 warnings. Nothing had looked at them, on the
reasoning that they were noise; four were not.

sg_solve.c let an uninitialised value decide a return status. ar_status
was declared without an initialiser, assigned only inside the phase-1 and
phase-2 blocks, and read at the end to choose between SG_STATUS_LIMIT and
SG_STATUS_OK. A run where neither phase executed reported whichever the
stack happened to hold. The second ar_status in the same file is left
alone: it is assigned before every read, so initialising it would only
have attached a comment that is untrue of it.

sg_concat.c skipped its memset on an early return. sg_seg_init_single()
checked !ctx || !stop || !seg and returned before zeroing the segment,
while its callers keep that segment on the stack and read it immediately
afterwards. The seg check now comes first, the memset follows, and the
other arguments are checked after it, so a non-NULL seg is always
written.

ralph_parse_lp.c called strncpy without terminating. Two sites copied
LP_MAX_NAME - 1 bytes into a name buffer and stopped; a name of exactly
that length leaves the buffer unterminated unless it happened to be
zeroed already. Both terminate explicitly now, which is what
ralph_api.c's equivalent already did.

cuts.c passed unvalidated counts to calloc. conflict_graph_create() took
num_sets and num_elements as int and used them directly; gcc reported the
argument range as [2^64 - 2^31, 2^64 - 1], having proven they could be
negative. calloc would have failed rather than overflowed, but failing
for an unreadable reason is worse than rejecting the input, so
non-positive counts return NULL.

  -Wmaybe-uninitialized     7 -> 0
  -Walloc-size-larger-than  4 -> 0
  -Wstringop-truncation     4 -> 2
  total                   136 -> 121

The two remaining stringop-truncation are correct code that terminates
explicitly, and gcc warning that truncation is possible at all.

Of what is left, 76 are miniz.h's static functions and 3 are TRE
fallthroughs. Of the 36 in OTTO code, 25 are -Wtype-limits on deliberate
32-bit overflow guards -- comparisons that are tautologically false on a
64-bit target and load-bearing on wasm32, already commented as such.
Silencing those would mean deleting a guard that matters or wrapping each
in a preprocessor conditional, so they stay as documented intent.

make all exits 0 and ralph, shared, surge, velo, nexus and arbor all pass
their suites.

+++ ralph/src/simplex.c
@@ -2537,7 +2537,7 @@ int simplex_pivot(SimplexTableau *tab,
                     goto basis_update_done;
                 }
                 lu_update_status = -2;
-                lu_reason = (tab->lu) ? lu_get_last_failure_reason(tab->lu) : lu_reason;
+                lu_reason = (tab->lu) ? (int)lu_get_last_failure_reason(tab->lu) : lu_reason;
                 growth_factor = (tab->lu) ? lu_get_growth_factor(tab->lu) : growth_factor;
                 lu_num_updates = (tab->lu) ? lu_get_num_updates(tab->lu) : lu_num_updates;
                 action = lp_refactor_policy_choose_basis_action(
@@ -2558,7 +2558,7 @@ int simplex_pivot(SimplexTableau *tab,
                     goto basis_update_done;
                 }
                 lu_update_status = -3;
-                lu_reason = (tab->lu) ? lu_get_last_failure_reason(tab->lu) : lu_reason;
+                lu_reason = (tab->lu) ? (int)lu_get_last_failure_reason(tab->lu) : lu_reason;
                 growth_factor = (tab->lu) ? lu_get_growth_factor(tab->lu) : growth_factor;
                 lu_num_updates = (tab->lu) ? lu_get_num_updates(tab->lu) : lu_num_updates;
                 action = lp_refactor_policy_choose_basis_action(
